+++ Tzkt.Data/Models/Scripts/Storage.cs
@@ -7,9 +7,9 @@ namespace Tzkt.Data.Models
         public int Id { get; set; }
         public int Level { get; set; }
         public int ContractId { get; set; }
-        public int? OriginationId { get; set; }
-        public int? TransactionId { get; set; }
-        public int? MigrationId { get; set; }
+        public long? OriginationId { get; set; }
+        public long? TransactionId { get; set; }
+        public long? MigrationId { get; set; }
         public bool Current { get; set; }
 
         public byte[] RawValue { get; set; }
